Added hierarchical removing Attribute Groups for Product Family and Product

+++ client/modules/pim/src/views/product/record/panels/product-attribute-values.js
@@ -92,6 +92,12 @@ Espo.define('pim:views/product/record/panels/product-attribute-values', ['views/
                 e.stopPropagation();
                 let data = $(e.currentTarget).data();
                 this.unlinkAttributeGroup(data);
+            },
+            'click [data-action="unlinkAttributeGroupHierarchy"]': function (e) {
+                e.preventDefault();
+                e.stopPropagation();
+                let data = $(e.currentTarget).data();
+                this.unlinkAttributeGroupHierarchy(data);
             }
         }, Dep.prototype.events),
 
@@ -683,6 +689,42 @@ Espo.define('pim:views/product/record/panels/product-attribute-values', ['views/
                     data: JSON.stringify({
                         ids: [this.model.id],
                         foreignIds: group.rowList
+                    }),
+                    type: 'DELETE',
+                    contentType: 'application/json',
+                    success: function () {
+                        this.notify('Unlinked', 'success');
+                        this.model.trigger('after:unrelate', this.link, this.defs);
+                        this.actionRefresh();
+                    }.bind(this),
+                    error: function () {
+                        this.notify('Error occurred', 'error');
+                    }.bind(this),
+                });
+            }, this);
+        },
+
+        unlinkAttributeGroupHierarchy(data) {
+            let id = data.id;
+            if (!id) {
+                return;
+            }
+
+            let group = this.groups.find(group => group.id === id);
+            if (!group || !group.rowList) {
+                return;
+            }
+
+            this.confirm({
+                message: this.translate('unlinkAttributeGroupConfirmation', 'messages', 'AttributeGroup'),
+                confirmText: this.translate('Unlink')
+            }, function () {
+                this.notify('Unlinking...');
+                $.ajax({
+                    url: `${this.scope}/action/unlinkAttributeGroupHierarchy`,
+                    data:  JSON.stringify({
+                        attributeGroupId: id,
+                        productId: this.model.id
                     }),
                     type: 'DELETE',
                     contentType: 'application/json',
